+++ src/Models/PageDefaults.php
@@ -44,6 +44,11 @@ trait PageDefaults
         return (new ResourceKeyFormat(static::class))->getKey();
     }
 
+    public function allowContexts(): bool
+    {
+        return true;
+    }
+
     public function allowMultipleContexts(): bool
     {
         return config('chief.allow_multiple_contexts', false);